pyproject.toml
[project]
name = "250728-polarh10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleak>=1.0.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "plotly-resampler>=0.11.0rc1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
    "streamlit-autorefresh>=1.0.1",
]
